requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
asgiref==3.9.1
    # via
    #   -r requirements/base.txt
    #   django
certifi==2025.8.3
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via
    #   code-annotations
    #   edx-django-utils
code-annotations==1.8.2
    # via -r requirements/test.in
coverage[toml]==7.10.5
    # via pytest-cov
cryptography==45.0.6
    # via pyjwt
django==4.2.23
    # via
    #   -c https:/raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
django-crum==0.7.9
    # via edx-django-utils
django-waffle==5.0.0
    # via
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.16.1
    # via
    #   -r requirements/test.in
    #   drf-jwt
    #   edx-drf-extensions
dnspython==2.7.0
    # via pymongo
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-django-utils==8.0.0
    # via
    #   -r requirements/test.in
    #   edx-drf-extensions
edx-drf-extensions==10.6.0
    # via -r requirements/test.in
edx-opaque-keys==3.0.0
    # via edx-drf-extensions
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
openedx-atlas==0.7.0
    # via -r requirements/base.txt
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psutil==7.0.0
    # via edx-django-utils
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.10.1
    # via
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.14.1
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pytest==8.4.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
requests==2.32.5
    # via edx-drf-extensions
semantic-version==2.10.0
    # via edx-drf-extensions
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.5.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
urllib3==2.5.0
    # via requests
